@@ -10,28 +10,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.0",
-    "@angular/cdk": "~16.1.2",
-    "@angular/common": "^16.0.0",
-    "@angular/compiler": "^16.0.0",
-    "@angular/core": "^16.0.0",
-    "@angular/forms": "^16.0.0",
-    "@angular/material": "~16.1.2",
-    "@angular/platform-browser": "^16.0.0",
-    "@angular/platform-browser-dynamic": "^16.0.0",
-    "@angular/router": "^16.0.0",
+    "@angular/animations": "^18.0.1",
+    "@angular/cdk": "~18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.1",
+    "@angular/material": "~18.0.1",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.1",
+    "@angular/router": "^18.0.1",
+    "@canvasjs/angular-charts": "^1.2.0",
     "apexcharts": "^3.52.0",
+    "date-fns-tz": "^3.1.3",
     "jwt-decode": "^4.0.0",
     "ng-apexcharts": "^1.11.0",
     "ng2-charts": "^6.0.1",
     "primeflex": "^3.3.1",
     "primeicons": "^6.0.1",
-    "primeng": "^16.0.0-rc.2"
+    "primeng": "17.18.9"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.0",
-    "@angular/cli": "~16.2.0",
-    "@angular/compiler-cli": "^16.2.0",
+    "@angular-devkit/build-angular": "^18.2.0",
+    "@angular/cli": "~18.2.0",
+    "@angular/compiler-cli": "^18.2.0",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
@@ -39,6 +41,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.1.3"
+    "typescript": "^5.4.0"
   }
 }
