@@ -1,3 +1,4 @@
+// menu.component.ts
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
@@ -7,10 +8,13 @@ import { Router } from '@angular/router';
   styleUrls: ['./menu.component.scss']
 })
 export class MenuComponent {
+  menuOpen = true;
 
-  constructor(
-    private router: Router,
-  ){}
+  constructor(private router: Router) {}
+
+  toggleMenu() {
+    this.menuOpen = !this.menuOpen;
+  }
 
   logout() {
     localStorage.removeItem('access_token');
