@@ -1,5 +1,7 @@
 export const environment = {
     production: false,
-    urlApi: 'https://estadodigital.fun/api/v1',
+    // urlApi: 'https://estadodigital.fun/api/v1',
+        urlApi: 'http://localhost:8000/api/v1',
+
 };
   
